bigfix on db dtype

+++ terra-update-batch/resources/params.py
@@ -282,8 +282,8 @@ DB_COLUMN_TYPE = {
         'FLOW': int,
         'cpa': str,
         'PERIOD': int,
-        'val_cpa': np.int64,
-        'q_kg': np.int64
+        'val_cpa': np.float64,
+        'q_kg': np.float64
         },
     'comext_exp': {
         'DECLARANT_ISO': str,
@@ -291,16 +291,16 @@ DB_COLUMN_TYPE = {
         'FLOW': int,
         'cpa': str,
         'PERIOD': int,
-        'val_cpa': np.int64,
-        'q_kg': np.int64
+        'val_cpa': np.float64,
+        'q_kg': np.float64
         },
     'cpa_intra': {
         'DECLARANT_ISO': str,
         'PARTNER_ISO': str,
         'FLOW': int,
         'PRODUCT': str,
-        'PERIOD': np.int64,
-        'VALUE_IN_EUROS': np.int64
+        'PERIOD': np.float64,
+        'VALUE_IN_EUROS': np.float64
         },
     'cpa_trim': {
         'DECLARANT_ISO': str,
@@ -308,8 +308,8 @@ DB_COLUMN_TYPE = {
         'FLOW': int,
         'cpa': str,
         'trimestre': str,
-        'val_cpa': np.int64,
-        'q_kg': np.int64
+        'val_cpa': np.float64,
+        'q_kg': np.float64
         },
     'tr_extra_ue': {
         'PRODUCT_NSTR': str,
@@ -318,8 +318,8 @@ DB_COLUMN_TYPE = {
         'PERIOD': int,
         'TRANSPORT_MODE': int,
         'FLOW': int,
-        'VALUE_IN_EUROS': np.int64,
-        'QUANTITY_IN_KG': np.int64
+        'VALUE_IN_EUROS': np.float64,
+        'QUANTITY_IN_KG': np.float64
         },
     'tr_extra_ue_trim': {
         'PRODUCT_NSTR': str,
@@ -328,7 +328,7 @@ DB_COLUMN_TYPE = {
         'TRIMESTRE': str,
         'TRANSPORT_MODE': int,
         'FLOW': int,
-        'VALUE_IN_EUROS': np.int64,
-        'QUANTITY_IN_KG': np.int64
+        'VALUE_IN_EUROS': np.float64,
+        'QUANTITY_IN_KG': np.float64
         },
 }
